+++ xilem_core/examples/user_interface.rs
@@ -42,17 +42,6 @@ impl Widget for Box<dyn Widget> {
 // due to the orphan rules...
 impl<W: Widget> ViewElement for WidgetPod<W> {
     type Mut<'a> = WidgetMut<'a, W>;
-
-    /* /// This implementation will perform `merge_up` multiple times, but that's
-    /// already true for downcasting anyway, so merge_up is already idempotent
-    fn with_reborrow_val<R: 'static>(
-        this: Self::Mut<'_>,
-        f: impl FnOnce(Self::Mut<'_>) -> R,
-    ) -> (Self::Mut<'_>, R) {
-        let value = WidgetMut { value: this.value };
-        let ret = f(value);
-        (this, ret)
-    } */
 }
 
 impl<State, Action> View<State, Action, ViewCtx> for Button {
